code coverage enabled, test.db moved to memory, small fixes in tests for in-memory databases
