+++ .gitignore
@@ -1,4 +1,5 @@
 /.idea/
 /vendor/
 /index.php
+/tests/test.db
 /composer.lock
