+++ phpunit.xml
@@ -16,10 +16,12 @@
          disallowTestOutput="true">
     <testsuites>
         <testsuite name="DrMVC ORM tests">
-            <directory suffix=".php">./test/</directory>
+            <directory suffix=".php">./tests/</directory>
         </testsuite>
     </testsuites>
-
+    <logging>
+        <log type="coverage-text" target="php://stdout" showUncoveredFiles="true"/>
+    </logging>
     <filter>
         <whitelist processUncoveredFilesFromWhitelist="true">
             <directory suffix=".php">src</directory>
