+++ src/Orm.php
@@ -2,12 +2,10 @@
 
 namespace DrMVC;
 
-use DrMVC\Orm\Orm as DrMVCOrm;
-
 /**
  * Class ORM - Abstract class for access to factory
  * @package DrMVC
  */
-class Orm extends DrMVCOrm
+class Orm extends Orm\Orm
 {
 }
